lxd/storage: Move storage_cgo.go to drivers package
