+++ lxd/storage_btrfs.go
@@ -25,6 +25,7 @@ import (
 	"github.com/lxc/lxd/lxd/rsync"
 	"github.com/lxc/lxd/lxd/state"
 	driver "github.com/lxc/lxd/lxd/storage"
+	"github.com/lxc/lxd/lxd/storage/drivers"
 	"github.com/lxc/lxd/lxd/util"
 	"github.com/lxc/lxd/shared"
 	"github.com/lxc/lxd/shared/api"
@@ -428,7 +429,7 @@ func (s *storageBtrfs) StoragePoolMount() (bool, error) {
 			// Since we mount the loop device LO_FLAGS_AUTOCLEAR is
 			// fine since the loop device will be kept around for as
 			// long as the mount exists.
-			loopF, loopErr := driver.PrepareLoopDev(source, driver.LoFlagsAutoclear)
+			loopF, loopErr := drivers.PrepareLoopDev(source, drivers.LoFlagsAutoclear)
 			if loopErr != nil {
 				return false, loopErr
 			}
